Improve compatability with modded labs and endgame techs
Some mods add specialised labs not intended to hold
every science pack. By forcibly adding the science pack
to every lab in `data-final-fixes.lua`, it only becomes
a matter of load order determining whether or not a
given lab would mistakenly include it. In addition,
some techs, like Frozeta's Science pack productivity,
want to have every science pack available to the default
lab, which may not have Tenebris's science pack if it
was loaded later.

Only adding the default lab instead of also the biolab is
fine due to the dependency on PlanetsLib, which mirrors all
default lab sciences into the biolab in `data-final-fixes.lua`.
It also has an option for mods to mark their own labs to
mirror the default lab sciences as well (or they might do
so manually), meaning that Tenebris doesn't have to worry
about that either.